adds a little animation when you hover on stuff
adds a little animation when you hover on stuff, or if you tap it on mobile, turns out it doesn't fw the text or anything. sidenote, the youtube thumbnails had a white background which meant that border-radius gave it a funny white border to the edge that looked like a bug, it might just be my shitty display but either way i fixed it.